Changed Version to 0.2.5.
This should be the last version before I realease it online, on
curse.com. All of the texture issues should be fixed, other than the
ones from the Trees, which I still need to fix and then finish. The
superhero texture selection has been added, and now it works. I still
need to add validation for the texture files, so that only valid
filenames will work, if you enter an option that is not present/valid,
it will just set it to default.