Started to remove Superhero Files.
I made a backup of the superhero related classes, so that I can remove
them from the actual mod. I also added language localizations for what
should be all of the Added Blocks. I did that the same way I added so
many superhero costumes; I wrote a program that generates the code!